requirements.txt
# Core
aiohttp>=3.8.0
websockets>=11.0
pyyaml>=6.0

# Audio
numpy>=1.20.0

# Web UI
jinja2>=3.0.0

# Optional: PipeWire integration
# pip install pulsectl  # For PulseAudio/PipeWire control

# Development
# pytest>=7.0.0
# pytest-asyncio>=0.20.0
